pyproject.toml
[tool.poetry]
name = "testapi"
version = "0.1.0"
description = ""
authors = ["leehjhjhj <tsukiakarii@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
requests = "^2.31.0"
python-decouple = "^3.8"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
